Organize all test components into testing mod
