+++ stellar_contract_sdk/src/testing/host/mem.rs
