+++ stellar_contract_sdk/src/testing/host/mod.rs
@@ -8,18 +8,8 @@
 
 pub mod fs;
 pub mod mem;
-use core::fmt::Debug;
 
 use crate::{Object, Val};
-
-impl Debug for Val {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Val")
-            .field("body", &self.get_body())
-            .field("tag", &self.get_tag())
-            .finish()
-    }
-}
 
 pub trait MockHost {
     fn log_value(&mut self, v: Val) -> Val;
