+++ stellar_contract_sdk/src/testing/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(not(target_family = "wasm"))]
+pub mod host;
