+++ stellar_contract_sdk/src/val.rs
@@ -302,3 +302,16 @@ impl Val {
         obj.into()
     }
 }
+
+#[cfg(not(target_family = "wasm"))]
+use core::fmt::Debug;
+
+#[cfg(not(target_family = "wasm"))]
+impl Debug for Val {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Val")
+            .field("body", &self.get_body())
+            .field("tag", &self.get_tag())
+            .finish()
+    }
+}
